Extract and refactor update-backup
